requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.4
djangorestframework==3.12.2
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
psycopg2-binary>=2.8
